+++ packages/core/src/data/position.ts
@@ -1,34 +1,20 @@
 import { allOf, juxtapose, oneOf } from '../combinators'
-import { keyword } from './keyword'
+import { keywords } from './keyword'
 import { lengthPercentage } from './length-percentage'
 
 export const position = oneOf([
-  keyword('left'),
-  keyword('center'),
-  keyword('right'),
-  keyword('top'),
-  keyword('bottom'),
+  keywords(['top', 'right', 'bottom', 'left', 'center']),
   lengthPercentage(),
   allOf([
-    oneOf([keyword('left'), keyword('center'), keyword('right')]),
-    oneOf([keyword('top'), keyword('center'), keyword('bottom')]),
+    keywords(['left', 'center', 'right']),
+    keywords(['top', 'center', 'bottom']),
   ]),
   juxtapose([
-    oneOf([
-      keyword('left'),
-      keyword('center'),
-      keyword('right'),
-      lengthPercentage(),
-    ]),
-    oneOf([
-      keyword('top'),
-      keyword('center'),
-      keyword('bottom'),
-      lengthPercentage(),
-    ]),
+    oneOf([keywords(['left', 'center', 'right']), lengthPercentage()]),
+    oneOf([keywords(['top', 'center', 'bottom']), lengthPercentage()]),
   ]),
   allOf([
-    juxtapose([oneOf([keyword('left'), keyword('right')]), lengthPercentage()]),
-    juxtapose([oneOf([keyword('top'), keyword('bottom')]), lengthPercentage()]),
+    juxtapose([keywords(['left', 'right']), lengthPercentage()]),
+    juxtapose([keywords(['top', 'bottom']), lengthPercentage()]),
   ]),
 ])